+++ dune/gdt/basefunctionset/pdelab.hh
@@ -30,8 +30,8 @@ namespace BaseFunctionSet {
 
 // forward, to be used in the traits and to allow for specialization
 template< class PdelabSpaceImp, class EntityImp,
-          class DomainFieldImp, int domainDim,
-          class RangeFieldImp, int rangeDim, int rangeDimCols = 1 >
+          class DomainFieldImp, size_t domainDim,
+          class RangeFieldImp, size_t rangeDim, size_t rangeDimCols = 1 >
 class PdelabWrapper
 {
   static_assert(Dune::AlwaysFalse< PdelabSpaceImp >::value, "Untested for arbitrary dimension!");
@@ -40,14 +40,14 @@ class PdelabWrapper
 
 // forward, to allow for specialization
 template< class PdelabSpaceImp, class EntityImp,
-          class DomainFieldImp, int domainDim,
-          class RangeFieldImp, int rangeDim, int rangeDimCols >
+          class DomainFieldImp, size_t domainDim,
+          class RangeFieldImp, size_t rangeDim, size_t rangeDimCols >
 class PdelabWrapperTraits;
 
 
 //! Specialization for dimRange = 1, dimRangeRows = 1
 template< class PdelabSpaceImp, class EntityImp,
-          class DomainFieldImp, int domainDim,
+          class DomainFieldImp, size_t domainDim,
           class RangeFieldImp >
 class PdelabWrapperTraits< PdelabSpaceImp, EntityImp, DomainFieldImp, domainDim, RangeFieldImp, 1, 1 >
 {
@@ -66,7 +66,7 @@ private:
 
 //! Specialization for dimRange = 1, dimRangeRows = 1
 template< class PdelabSpaceType, class EntityImp,
-          class DomainFieldImp, int domainDim,
+          class DomainFieldImp, size_t domainDim,
           class RangeFieldImp >
 class PdelabWrapper< PdelabSpaceType, EntityImp, DomainFieldImp, domainDim, RangeFieldImp, 1, 1 >
   : public BaseFunctionSetInterface< PdelabWrapperTraits< PdelabSpaceType, EntityImp,
@@ -156,8 +156,8 @@ private:
 
 // forward, to be used in the traits and to allow for specialization
 template< class PdelabSpaceImp, class EntityImp,
-          class DomainFieldImp, int domainDim,
-          class RangeFieldImp, int rangeDim, int rangeDimCols = 1 >
+          class DomainFieldImp, size_t domainDim,
+          class RangeFieldImp, size_t rangeDim, size_t rangeDimCols = 1 >
 class PiolaTransformedPdelabWrapper
 {
   static_assert(Dune::AlwaysFalse< PdelabSpaceImp >::value, "Untested for these dimensions!");
@@ -165,8 +165,8 @@ class PiolaTransformedPdelabWrapper
 
 
 template< class PdelabSpaceImp, class EntityImp,
-          class DomainFieldImp, int domainDim,
-          class RangeFieldImp, int rangeDim >
+          class DomainFieldImp, size_t domainDim,
+          class RangeFieldImp, size_t rangeDim >
 class PiolaTransformedPdelabWrapperTraits
 {
   static_assert(domainDim == rangeDim, "Untested!");
@@ -185,8 +185,8 @@ private:
 
 
 template< class PdelabSpaceType, class EntityImp,
-          class DomainFieldImp, int domainDim,
-          class RangeFieldImp, int rangeDim >
+          class DomainFieldImp, size_t domainDim,
+          class RangeFieldImp, size_t rangeDim >
 class PiolaTransformedPdelabWrapper< PdelabSpaceType, EntityImp, DomainFieldImp, domainDim, RangeFieldImp, rangeDim, 1 >
   : public BaseFunctionSetInterface< PiolaTransformedPdelabWrapperTraits< PdelabSpaceType, EntityImp,
                                                                        DomainFieldImp, domainDim,
@@ -311,8 +311,8 @@ private:
 
 
 template< class PdelabSpaceImp, class EntityImp,
-          class DomainFieldImp, int domainDim,
-          class RangeFieldImp, int rangeDim, int rangeDimCols = 1 >
+          class DomainFieldImp, size_t domainDim,
+          class RangeFieldImp, size_t rangeDim, size_t rangeDimCols = 1 >
 class PdelabWrapper
 {
   static_assert(AlwaysFalse< PdelabSpaceImp >::value, "You are missing dune-pdelab!");
@@ -320,8 +320,8 @@ class PdelabWrapper
 
 
 template< class PdelabSpaceImp, class EntityImp,
-          class DomainFieldImp, int domainDim,
-          class RangeFieldImp, int rangeDim, int rangeDimCols = 1 >
+          class DomainFieldImp, size_t domainDim,
+          class RangeFieldImp, size_t rangeDim, size_t rangeDimCols = 1 >
 class PiolaTransformedPdelabWrapper
 {
   static_assert(AlwaysFalse< PdelabSpaceImp >::value, "You are missing dune-pdelab!");
