+++ dune/gdt/playground/localevaluation/OS2014.hh
@@ -127,7 +127,7 @@ public:
                            diffusive_flux_.local_function(entity));
   }
 
-  template< class R, int rT, int rCT, int rA, int rCA >
+  template< class R, size_t rT, size_t rCT, size_t rA, size_t rCA >
   size_t order(const LocalfunctionTupleType& localFuncs,
                const Stuff::LocalfunctionSetInterface
                    < EntityType, DomainFieldType, dimDomain, R, rT, rCT >& testBase,
@@ -146,7 +146,7 @@ public:
   } // ... order(...)
 
 private:
-  template< class R, int rLD, int rCLD, int rLDT, int rCLDT, int rLDF, int rCLDF, int rT, int rCT, int rA, int rCA >
+  template< class R, size_t rLD, size_t rCLD, size_t rLDT, size_t rCLDT, size_t rLDF, size_t rCLDF, size_t rT, size_t rCT, size_t rA, size_t rCA >
   size_t redirect_order(const Stuff::LocalfunctionInterface
                             < EntityType, DomainFieldType, dimDomain, R, rLD, rCLD >& local_diffusion_factor,
                         const Stuff::LocalfunctionInterface
@@ -172,7 +172,7 @@ private:
   } // ... redirect_order(...)
 
 public:
-  template< class R, int rT, int rCT, int rA, int rCA >
+  template< class R, size_t rT, size_t rCT, size_t rA, size_t rCA >
   void evaluate(const LocalfunctionTupleType& localFuncs,
                 const Stuff::LocalfunctionSetInterface
                     < EntityType, DomainFieldType, dimDomain, R, rT, rCT >& test_base,
@@ -193,7 +193,7 @@ public:
   } // ... evaluate(...)
 
 private:
-  template< class R, int rLD, int rCLD, int rLDh, int rCLDh, int rLDT, int rCLDT, int rLDF, int rCLDF, int rT, int rCT, int rA, int rCA >
+  template< class R, size_t rLD, size_t rCLD, size_t rLDh, size_t rCLDh, size_t rLDT, size_t rCLDT, size_t rLDF, size_t rCLDF, size_t rT, size_t rCT, size_t rA, size_t rCA >
   void redirect_evaluate(const Stuff::LocalfunctionInterface
                              < EntityType, DomainFieldType, dimDomain, R, rLD, rCLD >& /*local_diffusion_factor*/,
                          const Stuff::LocalfunctionInterface
